api/requirements.txt
asgiref==3.2.5
backcall==0.1.0
decorator==4.0.6
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.12.2
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.1.0
jedi==0.16.0
parso==0.6.2
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.2.0
prompt-toolkit==3.0.4
ptyprocess==0.5
Pygments==2.6.1
python-dateutil==2.6.0
pytz==2019.3
reportlab==3.3.0
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.8
gunicorn==20.1.0
uritemplate==3.0.1
PyYAML==5.4.1
django_filter==2.4.0
model-bakery==1.3.2
